requirements.txt
binaryornot==0.4.4
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.4
cycler==0.12.1
decorator==4.4.2
Flask==3.0.1
Flask-JWT==0.3.2
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.3
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
mysql-connector-python==8.3.0
numpy==1.19.2
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
proglog==0.1.10
protobuf==4.25.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydub==0.25.1
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
rubicon-objc==0.4.7
six==1.16.0
SQLAlchemy==2.0.27
sympy==1.12
tqdm==4.66.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
